Add showFocusedBorder parameter to custom text field so that the focused border can be hidden on the control bar.
Could have gone with a separate chat bar here for just the control bar, because I'm not sure that we will use this option anywhere else. However, pending a more comprehensive overhaul of the CustomTextField / TextField transition, I didn't want to introduce a new "special case" for text fields here.